Initial change for #1o and #11 - rename and prep for PyPi publishing
Add correct license file, along with trove classifier.
Rename package along with source code directory.
